refactor: fixing import after change result class place
